requirements_voicebot.txt
pyaudio
google-cloud-speech
google-cloud-texttospeech
requests
